+++ database/migrations/2023_05_24_130531_create_realestat_bookings_table.php
@@ -13,30 +13,31 @@ class CreateRealestatBookingsTable extends Migration
      */
     public function up()
     {
-        Schema::create('realestat_bookings', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('request_id');
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('realestate_id');
-            $table->unsignedBigInteger('booking_type');
-            $table->timestamps();
+        if (!Schema::hasTable('realestat_bookings')) {
+            Schema::create('realestat_bookings', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('request_id');
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('realestate_id');
+                $table->unsignedBigInteger('booking_type');
+                $table->timestamps();
 
-            $table->foreign('request_id')
-            ->references('id')
-            ->on('requests')
-            ->onDelete('cascade');
+                $table->foreign('request_id')
+                    ->references('id')
+                    ->on('requests')
+                    ->onDelete('cascade');
 
-            $table->foreign('realestate_id')
-            ->references('id')
-            ->on('real_states')
-            ->onDelete('cascade');
-            
-            $table->foreign('booking_type')
-            ->references('id')
-            ->on('categories')
-            ->onDelete('cascade');
-            
-        });
+                $table->foreign('realestate_id')
+                    ->references('id')
+                    ->on('real_states')
+                    ->onDelete('cascade');
+
+                $table->foreign('booking_type')
+                    ->references('id')
+                    ->on('categories')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
